backend/requirements.txt
fastapi
groq
elevenlabs
gtts
speechrecognition
pydub
pyaudio
uvicorn[standard]
python-dotenv
python-multipart